Add -p option to perform parallel ping testing
This commits adds to LTNT_test_manager the -p (or --add-ping)
option, to perform, together and in parallel with the LaTe tests,
also a ping RTT test.
This option is disabled by default (i.e. ping is not normally
used unless explicitely requested).
This option currently affects only the master, as it will be
the master that will launch ping towards the DUT IP address
of the slave.
This may be changed in the future, allowing also the slave to
start a ping towards the master.