Introducing LaTe and iPerf path options in LTNT.ini
This commit introduces two new configuration options in LTNT.ini:
exec_path_late and exec_path_iperf.
They can be used to set a non-standard path for LaTe and iPerf,
when not using LTNT with OpenWrt (for instance, with Ubuntu it may
be less convienent to place the LaTe executable inside /root, which
is the default expected path where to look for LaTe).
These options currently affect both the master and the slave device
participating in a measurement session.
This commit also improves the behaviour of LTNT Test Manager when
cleaning and creating the directories for the logs.