Introducing options to set the -b and -P iPerf argument
This commit introduces three new configuration options, which can
be set with the LTNT.ini configuration file and which can be used
to tune the iPerf UDP and TCP offered traffic (-b option) and the
number of UDP client threads to be started for each UDP test (-P
option).
It also updates README.md accordingly.